created a parent player div which will contain all detail components
