+++ nba/src/components/Player/Player.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import PlayerName from './PlayerName';
+
+const Player = () => {
+  return (
+    <div>
+      <PlayerName />
+    </div>
+  )
+
+}
+
+export default Player;
